refactor: move ITextMessageService to domain, fix code review feedback
- Move ITextMessageService interface to Shadowbrook.Domain/Common (DDD pattern)
- Convert static methods to instance methods, inject via DI
- Fix Results.Created() to Results.Ok() on inbound endpoint
- Add this. qualification per editorconfig rules
