+++ src/backend/Shadowbrook.Api/Infrastructure/Services/ConsoleTextMessageService.cs
@@ -1,3 +1,5 @@
+using Shadowbrook.Domain.Common;
+
 namespace Shadowbrook.Api.Infrastructure.Services;
 
 /// <summary>
