+++ src/backend/Shadowbrook.Api/Infrastructure/Services/InMemoryTextMessageService.cs
@@ -1,4 +1,5 @@
 using System.Collections.Concurrent;
+using Shadowbrook.Domain.Common;
 
 namespace Shadowbrook.Api.Infrastructure.Services;
 
@@ -15,31 +16,31 @@ public enum SmsDirection
 /// </summary>
 public class InMemoryTextMessageService(ILogger<InMemoryTextMessageService> logger) : ITextMessageService
 {
-    private static readonly ConcurrentBag<SmsMessage> messages = [];
+    private readonly ConcurrentBag<SmsMessage> messages = [];
 
     public const string SystemPhoneNumber = "+10000000000";
 
     public Task SendAsync(string toPhoneNumber, string message, CancellationToken cancellationToken = default)
     {
         var sms = new SmsMessage(SystemPhoneNumber, toPhoneNumber, message, DateTimeOffset.UtcNow, SmsDirection.Outbound);
-        messages.Add(sms);
+        this.messages.Add(sms);
         logger.LogInformation("[SMS] To: {PhoneNumber} | Message: {Message}", toPhoneNumber, message);
         return Task.CompletedTask;
     }
 
-    public static void AddInbound(string fromPhoneNumber, string message)
+    public void AddInbound(string fromPhoneNumber, string message)
     {
         var sms = new SmsMessage(fromPhoneNumber, SystemPhoneNumber, message, DateTimeOffset.UtcNow, SmsDirection.Inbound);
-        messages.Add(sms);
+        this.messages.Add(sms);
     }
 
-    public static IReadOnlyList<SmsMessage> GetAll() =>
-        messages.OrderByDescending(m => m.Timestamp).ToList();
+    public IReadOnlyList<SmsMessage> GetAll() =>
+        this.messages.OrderByDescending(m => m.Timestamp).ToList();
 
-    public static IReadOnlyList<SmsMessage> GetByPhone(string phoneNumber) =>
-        messages.Where(m => m.From == phoneNumber || m.To == phoneNumber)
+    public IReadOnlyList<SmsMessage> GetByPhone(string phoneNumber) =>
+        this.messages.Where(m => m.From == phoneNumber || m.To == phoneNumber)
                 .OrderBy(m => m.Timestamp)
                 .ToList();
 
-    public static void Clear() => messages.Clear();
+    public void Clear() => this.messages.Clear();
 }
